# frontend/openapi.json
{
  "components": {
    "responses": {
      "DEFAULT_ERROR": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Default error response"
      },
      "UNPROCESSABLE_ENTITY": {
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        },
        "description": "Unprocessable Entity"
      }
    },
    "schemas": {
...
      "ConvertersListResponse": {
        "additionalProperties": false,
        "properties": {
          "converters": {
            "items": {
              "additionalProperties": {},
              "type": "object"
            },
            "type": "array"
          }
        },
        "required": [
          "converters"
        ],
        "type": "object"
      },
...
      "DatasetClearResponse": {
        "additionalProperties": false,
        "properties": {
          "ok": {
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "DatasetLoadDemoRequest": {
        "additionalProperties": false,
        "properties": {
          "clipper": {
            "default": "",
            "type": "string"
          },
          "converter": {
            "default": "",
            "type": "string"
          },
          "dataset_name": {
            "default": "",
            "type": "string"
          },
          "embedder": {
            "default": "",
            "type": "string"
          },
          "name": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "type": "object"
      },
      "DatasetLoadFolderRequest": {
        "additionalProperties": false,
        "properties": {
          "media_type": {
            "default": "audio",
            "type": "string"
          },
          "path": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "path"
        ],
        "type": "object"
      },
      "DatasetLoadSourceRequest": {
        "additionalProperties": false,
        "properties": {
          "source": {
            "additionalProperties": {},
            "description": "Origin dict as stored on medias.",
            "type": "object"
          }
        },
        "required": [
          "source"
        ],
        "type": "object"
      },
      "DatasetLoadStartedResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          },
          "task_id": {
            "type": "string"
          }
        },
        "required": [
          "message",
          "ok",
          "task_id"
        ],
        "type": "object"
      },
      "DatasetRegistryLoadResponse": {
        "additionalProperties": false,
        "properties": {
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          },
          "task_id": {
            "default": "",
            "type": "string"
          }
        },
        "required": [
          "message",
          "ok"
        ],
        "type": "object"
      },
...
      "DetectorRegistryLoadRequest": {
        "additionalProperties": false,
        "properties": {
          "detector_id": {
            "default": null,
            "nullable": true,
            "type": "string"
          }
        },
        "type": "object"
      },
      "DetectorRegistryLoadResponse": {
        "additionalProperties": false,
        "properties": {
          "examples_seeded": {
            "type": "integer"
          },
          "labels_restored": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "ok": {
            "type": "boolean"
          },
          "task_id": {
            "type": "string"
          }
        },
        "required": [
          "ok"
        ],
        "type": "object"
      },
...
      "Error": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "description": "Error code",
            "type": "integer"
          },
          "errors": {
            "additionalProperties": {},
            "description": "Errors",
            "type": "object"
          },
          "message": {
            "description": "Error message",
            "type": "string"
          },
          "status": {
            "description": "Error name",
            "type": "string"
          }
        },
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "VTSearch API",
    "version": "snapshot"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/converters": {
      "get": {
        "description": "Query parameters:\ntarget: A ``type_id`` (e.g. ``\"image\"``) or ``folder_import_name``\n    (e.g. ``\"images\"``).  When provided, only converters whose\n    ``target_type`` matches are returned.\nsource: A ``type_id`` (e.g. ``\"video\"``) or ``folder_import_name``\n    (e.g. ``\"videos\"``).  When provided, only converters whose\n    ``source_type`` matches are returned.",
        "parameters": [
          {
            "in": "query",
            "name": "target",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConvertersListResponse"
                }
              }
            },
            "description": "OK"
          },
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Return all converters, optionally filtered by source or target media type.",
        "tags": [
          "datasets_listings"
        ]
      }
    },
...
    "/api/dataset/clear": {
      "post": {
        "description": "Uses the ``X-Dataset-Id`` header (via ``get_active_context()``) to\ndetermine which dataset to clear.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetClearResponse"
                }
              }
            },
            "description": "OK"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Clear the request-scoped dataset from memory.",
        "tags": [
          "datasets_load"
        ]
      }
    },
...
    "/api/dataset/export": {
      "get": {
        "description": "Success returns a binary ``application/octet-stream`` download; that\nbody is left undescribed in the spec (mirroring the audio / video /\nimage streaming routes in ``media/list.py``).",
        "responses": {
          "400": {
            "description": "No dataset is currently loaded."
          },
          "500": {
            "description": "Dataset export failed unexpectedly."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Export the current dataset to a pickle file.",
        "tags": [
          "datasets_load"
        ]
      }
    },
    "/api/dataset/import-local-folder": {
      "post": {
        "description": "The browser uses ``<input type=\"file\" webkitdirectory>`` to let the\nuser pick a directory; each selected ``File`` is appended to the\nmultipart body under the key ``\"files\"`` with its ``webkitRelativePath``\nas the multipart filename.  We stream each file to a temporary\ndirectory on the server (preserving sub-directory structure) and then\ndelegate to the regular folder importer to do the actual scanning,\nembedding, and dataset registration.  The temp directory is removed\nonce the importer finishes (success or failure).\n\nLocal-files uploads may additionally include a ``vectors_file`` form\nfield carrying a ``.npz`` archive of pre-computed embedding vectors\nkeyed by uploaded-file name (basename or relative path).  Files\nwhose name matches an NPZ key reuse the supplied vector instead of\nrunning the embedding model.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetLoadStartedResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Multipart body is missing the ``files`` field, has no valid files, is missing ``media_type``, or carries malformed ``clipper_params`` / ``vectors_file``."
          },
          "500": {
            "description": "The server_folder importer is unavailable."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Import a folder uploaded from the user's *browser* machine.",
        "tags": [
          "datasets_load"
        ]
      }
    },
...
    "/api/dataset/load-demo": {
      "post": {
        "description": "When a ``converter`` is specified, the demo data is loaded using its\noriginal media type, then converted to the converter's target type.\nThe resulting dataset has the *target* type, not the demo's original\ntype.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetLoadDemoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetLoadStartedResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Unknown demo dataset name."
          },
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "500": {
            "description": "The demo importer is unavailable."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Load a demo dataset in a background thread.",
        "tags": [
          "datasets_load"
        ]
      }
    },
    "/api/dataset/load-file": {
      "post": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetLoadStartedResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Multipart body is missing a file or no file is selected."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Load a dataset from an uploaded pickle file.",
        "tags": [
          "datasets_load"
        ]
      }
    },
    "/api/dataset/load-folder": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetLoadFolderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetLoadStartedResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Invalid or missing folder path."
          },
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Generate dataset from a folder of media files.",
        "tags": [
          "datasets_load"
        ]
      }
    },
    "/api/dataset/load-source": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DatasetLoadSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetLoadStartedResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Unknown importer, unreloadable origin, or invalid path inside origin."
          },
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Reload a dataset from a stored source origin dict.",
        "tags": [
          "datasets_load"
        ]
      }
    },
...
    "/api/datasets/registry/{dataset_id}/load": {
      "parameters": [
        {
          "in": "path",
          "name": "dataset_id",
          "required": true,
          "schema": {
            "minLength": 1,
            "type": "string"
          }
        }
      ],
      "post": {
        "description": "If the dataset is already loaded in memory, it is simply activated\n(made the current UI-facing dataset) without re-reading the pkl.",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetRegistryLoadResponse"
                }
              }
            },
            "description": "OK"
          },
          "403": {
            "description": "Access denied for the current user."
          },
          "404": {
            "description": "Dataset not found, or saved pkl file is missing."
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Load a registered dataset from its saved pkl file.",
        "tags": [
          "datasets_registry"
        ]
      }
    },
...
    "/api/detectors/registry/load": {
      "post": {
        "description": "Pass ``detector_id=null`` (or omit the field) to unload the active\ndetector without loading another one.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DetectorRegistryLoadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetectorRegistryLoadResponse"
                }
              }
            },
            "description": "OK"
          },
          "404": {
            "description": "Detector not found."
          },
          "422": {
            "$ref": "#/components/responses/UNPROCESSABLE_ENTITY"
          },
          "default": {
            "$ref": "#/components/responses/DEFAULT_ERROR"
          }
        },
        "summary": "Load a detector into memory and make it active.",
        "tags": [
          "detectors_registry"
        ]
      }
    },
...
  },
...
}